Move setup dialog code out of decodableReaderToolboxPanel to own file
